bug-hunt #76: -Wcast-qual — fix discarding const/volatile qualifiers
Fixes 143 -Wcast-qual violations across 50 source files:
- MMIO read accessors (io.h, vga.c, e1000.c, intel_gpu.c, usb_ehci.c)
- Block device write paths (blockdev.h, ahci.c, swap.c, bufcache.c, virtio_blk.c)
- ERR_CAST and ERR_PTR_OR_NULL helpers (err.h)
- Kernel spawn and process creation (spawn_kernel.c, syscall.c)
- Lockdep diagnostic prints (lockdep.c)
- Module decompression (module_compress.c)
- ACPI resource parsing (acpi.c, acpi_aml.c)
- EDID parsing (edid.c)
- ACPI CRS buffer parsing (acpi.c)
- Netlink attribute access (netlink.h, nla_data, NLA_NEXT)
- Network stack handler signatures (net.c, net_tcp.c, net_udp.c, ipv6*.c, etc.)
- Netfilter nf_tables eval expr casting (nf_tables.c)
- Socket setsockopt optval reads (socket.c)
- String library (string.c) - strchr, strstr, strcmp patterns
- bsearch in stdlib (stdlib.c)
- Various other const/volatile cleanups

-Wcast-qual flag is now permanently enabled in CFLAGS.

+++ src/drivers/acpi_aml.c
@@ -364,7 +364,7 @@ static int aml_add_node(const char name[4], uint8_t type, uint16_t parent,
 	node->parent = parent;
 	node->first_child = 0xFFFF;
 	node->next_sibling = 0xFFFF;
-	node->aml_start = (uint8_t *)aml_start;
+	node->aml_start = (uint8_t *)(uintptr_t)aml_start;
 	node->aml_length = aml_length;
 	node->from_ssdt = ssdt_index;
 	node->value = NULL;
